refactor(frontend): flatten veryl crate modules
