+++ crates/celox-frontend-veryl/src/dynamic_for_check.rs
