+++ crates/celox-frontend-veryl/src/hierarchy.rs
