+++ crates/celox-frontend-veryl/src/loop_provenance.rs
