+++ crates/celox-frontend-veryl/src/lowering/bitaccess.rs
