+++ crates/celox-frontend-veryl/src/lowering/bitslicer.rs
