+++ crates/celox-frontend-veryl/src/lowering/case.rs
