+++ crates/celox-frontend-veryl/src/lowering/context_width.rs
