+++ crates/celox-frontend-veryl/src/lowering/global_ff.rs
