+++ crates/celox-frontend-veryl/src/lowering/logic_tree/effect.rs
