+++ crates/celox-frontend-veryl/src/lowering/logic_tree/state.rs
