+++ crates/celox-frontend-veryl/src/lowering/registry.rs
